ci test: skip upgrade tests on release
Upgrade tests could fail in due to the existed package's external
dependency updates (e.g. new versions of MariaDB/MySQL) that do not
affect the functionality of the release package. And also, we have
confirmed that package upgrades work correctly in non-release
testing environments.

To avoid blocking new package releases due to irrelevant upgrade
test failures, we skip upgrade test on only release.

+++ packages/apt/test.sh
@@ -3,6 +3,7 @@
 set -exu
 
 package=$1
+triggered_branch=$2
 
 echo "::group::Prepare repository"
 
@@ -220,6 +221,12 @@ echo "::endgroup::"
 
 
 echo "::group::Upgrade test"
+if [[ "${triggered_branch}" =~ ^refs/tags/ ]]; then
+  echo "Skipping upgrade tests on release."
+  echo "::endgroup::"
+  exit 0
+fi
+
 sudo apt purge -V -y \
   ${package} \
   "${mysql_package_prefix}-*"
